+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # @koopjs/geoservice-utils
 
+## Unreleased
+### Major Changes
+- breaking change, rename "normalize" to "standardize"
+
 ## 1.0.0
 ### Major Changes
 - initial commit